line and point refactoring complete
