+++ geometry/__init__.py
@@ -1,0 +1,2 @@
+from .point import Point
+from .line import Line
