+++ geometry/utils.py
@@ -1,9 +1,4 @@
-try:
-    import Point
-except:
-    pass
-
-def distanceBetweenPoints(point1: Point, point2: Point) -> float:
+def distanceBetweenPoints(point1, point2) -> float:
     x1 = point1.xcoord
     y1 = point1.ycoord
     x2 = point2.xcoord
@@ -11,3 +6,16 @@ def distanceBetweenPoints(point1: Point, point2: Point) -> float:
     dist: float
     dist = ((x1 - x2)**2 + (y1-y2)**2)**0.5
     return dist
+
+
+def intersectionPointofTwoLines(l1, l2):
+    a, b, c = l1.xcoeff, l1.ycoeff, l1.constant
+    d, e, f = l2.xcoeff, l2.ycoeff, l2.constant
+    assert b/a != e/d, "Lines are parallel, no intersection"
+    k = b/e
+    l = a/d
+    x = (f*k-c)/(a-d*k)
+    y = (l*f-c)/(b-l*e)
+    poi = (x, y)
+    from geometry import Point
+    return Point(poi)
